rework tree dump
