+++ zenoblend/execute_operator.py
@@ -2,6 +2,7 @@ import bpy
 import time
 from . import scenario
 
+'''
 class ZenoApplyOperator(bpy.types.Operator):
     """Apply the Zeno graph"""
     bl_idname = "node.zeno_apply"
@@ -21,6 +22,7 @@ class ZenoApplyOperator(bpy.types.Operator):
             dt = time.time() - t0
             self.report({'INFO'}, 'Node tree applied in {:.04f}s'.format(dt))
         return {'FINISHED'}
+'''
 
 
 # class ZenoStopOperator(bpy.types.Operator):
@@ -61,7 +63,7 @@ class ZenoReloadOperator(bpy.types.Operator):
 def draw_menu(self, context):
     if context.area.ui_type == 'ZenoNodeTree':
         self.layout.separator()
-        self.layout.operator("node.zeno_apply", text="Apply Graph")
+        #self.layout.operator("node.zeno_apply", text="Apply Graph")
         #self.layout.operator("node.zeno_stop", text="Stop Running Graph")  # deprecated
         self.layout.operator("node.zeno_reload", text="Reload Graph Nodes")
 
@@ -80,8 +82,9 @@ class ZenoSceneProperties(bpy.types.PropertyGroup):
     ui_list_selected_tree: bpy.props.IntProperty(update=update_node_tree_list)
    
 
-class ZenoNewIndex():
+class ZenoNewIndex:
     new_index = -1
+
 
 class ZENO_UL_TreePropertyList(bpy.types.UIList):
     """Show in node tree editor"""
@@ -106,7 +109,7 @@ class ZENO_UL_TreePropertyList(bpy.types.UIList):
         row = row.row(align=True)
         row.alignment = 'RIGHT'
         row.ui_units_x = 3
-        row.prop(tree, 'zeno_enabled', icon= f"RESTRICT_VIEW_{'OFF' if tree.zeno_enabled else 'ON'}", text='', emboss=False)
+        row.prop(tree, 'zeno_enabled', icon='RESTRICT_VIEW_' + ('OFF' if tree.zeno_enabled else 'ON'), text='')#, emboss=False)
         row.prop(tree, 'zeno_realtime_update', icon='FILE_REFRESH', text='')
         row.prop(tree, 'zeno_cached', icon='PHYSICS', text='')
 
@@ -161,8 +164,9 @@ class ZenoScenePanel(bpy.types.Panel):
         if tree_id >= 0:
             tree = bpy.data.node_groups[tree_name_dict[tree_id]]           
             if tree.zeno_cached:
-                cached_to_frame = tree.nextFrameId - 1 if getattr(tree, "nextFrameId", None) else ''
+                cached_to_frame = tree._zeno_nextFrameId - 1 if getattr(tree, "_zeno_nextFrameId", None) else ''
                 col.label(text=f"Cached to frame: {cached_to_frame}")
+        '''
         row = layout.row()
         row.operator('node.zeno_apply')
         if tree.zeno_cached:
@@ -170,10 +174,11 @@ class ZenoScenePanel(bpy.types.Panel):
                 row.enabled = False # gray out button
         elif tree.zeno_realtime_update:
             row.enabled = False
+        '''
         
 
 classes = (
-    ZenoApplyOperator,
+    # ZenoApplyOperator,
     # ZenoStopOperator,
     ZenoReloadOperator,
     ZenoSceneProperties,
@@ -298,94 +303,3 @@ def load_from_zsg(prog):
 
 #bpy.load_zsg = lambda path: load_from_zsg(__import__('json').load(open(path)))
 '''
-
-
-def find_tree_sub_category(tree):
-    assert tree.bl_idname == 'ZenoNodeTree', tree
-    if 'SubCategory' in tree.nodes:
-        node = tree.nodes['SubCategory']
-        if node.zeno_type == 'SubCategory':
-            return node.inputs['name:'].default_value
-    for node_name, node in tree.nodes.items():
-        if not hasattr(node, 'zeno_type'): continue
-        if node.zeno_type == 'SubCategory':
-            return node.inputs['name:'].default_value
-    return 'uncategorized'
-
-
-def find_tree_sub_io_names(tree):
-    assert tree.bl_idname == 'ZenoNodeTree', tree
-    inputs = []
-    outputs = []
-    for node_name, node in tree.nodes.items():
-        if not hasattr(node, 'zeno_type'): continue
-        if node.zeno_type == 'SubInput':
-            type = node.inputs['type:'].default_value
-            name = node.inputs['name:'].default_value
-            defl = node.inputs['defl:'].default_value
-            inputs.append((type, name, defl))
-        elif node.zeno_type == 'SubOutput':
-            type = node.inputs['type:'].default_value
-            name = node.inputs['name:'].default_value
-            defl = node.inputs['defl:'].default_value
-            outputs.append((type, name, defl))
-    return inputs, outputs
-
-
-eval_bpy_data = {
-    # possibly support more bpy datablocks, like objects, images, textures
-    'texts': lambda data: data.as_string(),
-    'objects': lambda data: data.name,
-}
-
-def dump_tree(tree):
-    assert tree.bl_idname == 'ZenoNodeTree', tree
-    for node_name, node in tree.nodes.items():
-        if not hasattr(node, 'zeno_type'): continue
-        node_type = node.zeno_type
-        yield ('addNode', node_type, node_name)
-
-        # thank @hooyuser for contribute!
-        if hasattr(node, 'bpy_data_inputs'):
-            for input_name, data_type in node.bpy_data_inputs.items():
-                data_blocks = getattr(bpy.data, data_type)
-                data_block_name = getattr(node, input_name)
-                if data_block_name not in data_blocks:
-                    print('WARNING: object named `{}` not exist!')
-                    continue
-                data = data_blocks[data_block_name]
-                value = eval_bpy_data[data_type](data)
-                yield ('setNodeInput', node_name, input_name, value)
-
-        for input_name, input in node.inputs.items():
-            if input.is_linked:
-                if len(input.links) == 1:
-                    link = input.links[0]
-                    src_node_name = link.from_node.name
-                    src_socket_name = link.from_socket.name
-                    yield ('bindNodeInput', node_name, input_name,
-                            src_node_name, src_socket_name)
-            elif hasattr(input, 'default_value'):
-                value = input.default_value
-                if type(value).__name__ in ['bpy_prop_array', 'Vector']:
-                    value = tuple(value)
-                yield ('setNodeInput', node_name, input_name, value)
-
-        if node.zeno_type == 'Subgraph':
-            yield ('setNodeInput', node_name, 'name:', node.graph_name)
-        yield ('completeNode', node_name)
-
-
-def dump_all_trees():
-    yield ('clearAllState',)
-    for name, tree in bpy.data.node_groups.items():
-        if tree.bl_idname != 'ZenoNodeTree': continue
-        yield ('switchGraph', name)
-        yield from dump_tree(tree)
-
-
-def dump_scene():
-    import json
-    data = list(dump_all_trees())
-    data = json.dumps(data)
-    return data
